一些管理者用的prototype
