+++ src/main/java/tw/edu/ncu/cc/manage/controller/godmode/ClientManageController.java
@@ -1,0 +1,10 @@
+package tw.edu.ncu.cc.manage.controller.godmode;
+
+/**
+ * 應用服務管理
+ * @author yyc1217
+ *
+ */
+public class ClientManageController {
+
+}
