+++ src/main/java/tw/edu/ncu/cc/manage/controller/godmode/StatisticController.java
@@ -1,0 +1,10 @@
+package tw.edu.ncu.cc.manage.controller.godmode;
+
+/**
+ * 統計報表
+ * @author yyc1217
+ *
+ */
+public class StatisticController {
+
+}
