+++ src/main/java/tw/edu/ncu/cc/manage/controller/godmode/UserManageController.java
@@ -1,0 +1,10 @@
+package tw.edu.ncu.cc.manage.controller.godmode;
+
+/**
+ * 使用者管理
+ * @author yyc1217
+ *
+ */
+public class UserManageController {
+
+}
