+++ src/main/webapp/WEB-INF/layouts/components/leftsider.jsp
@@ -51,14 +51,18 @@
 	        <li class="treeview active">
 	            <a href="#">
 	                <i class="fa fa-unlock-alt"></i>
-	                <span>黑名單管理</span>
+	                <span>管理者功能</span>
 	                <i class="fa fa-angle-left pull-right"></i>
 	            </a>
 	            <ul class="treeview-menu">
-	                <li><a href="<c:url value="/blacklist"/>"><i class="fa fa-angle-double-right"></i> 黑名單管理</a></li>                                 
+                    <li><a href="<c:url value="/godmode/user"/>"><i class="fa fa-angle-double-right"></i>使用者授權管理</a></li>
+	                <li><a href="<c:url value="/godmode/client"/>"><i class="fa fa-angle-double-right"></i>應用服務管理</a></li>
+                    <li><a href="<c:url value="/godmode/blacklist/user"/>"><i class="fa fa-angle-double-right"></i>使用者黑名單管理</a></li>
+                    <li><a href="<c:url value="/godmode/blacklist/client"/>"><i class="fa fa-angle-double-right"></i>應用服務黑名單管理</a></li>
+                    <li><a href="<c:url value="/godmode/statistic"/>"><i class="fa fa-angle-double-right"></i>統計報表</a></li>
+                    <li><a href="<c:url value="/godmode/manager"/>"><i class="fa fa-angle-double-right"></i>系統管理者管理</a></li>
 	            </ul>
 	        </li>
-	
 	    </ul>
 	</section>
 	<%-- /.sidebar --%>
