+++ src/main/webapp/WEB-INF/views/godmode/manager/create.jsp
@@ -1,0 +1,26 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"%>
+<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
+<%@ taglib prefix="form" uri="http://www.springframework.org/tags/form"%>
+
+<%-- left column --%>
+<div class="box box-primary invoice">
+	<div class="box-header">
+		<h3 class="box-title">新增系統管理員</h3>
+	</div>
+	<%-- /.box-header --%>
+	<%-- form start --%>
+	<form:form role="form" action='#' method='post'>
+		<div class="box-body">
+			<div class="form-group">
+				<label for="account">系統管理者帳號</label>
+				<input id=account type="text" class="form-control" name="account">
+			</div>
+		</div>
+		<%-- /.box-body --%>
+
+		<div class="box-footer">
+			<button type="submit" class="btn btn-success">送出</button>
+		</div>
+	</form:form>
+</div>
+<%-- /.box --%>
